+++ host/Android.bp
@@ -68,7 +68,6 @@ gfxstream_backend_static_deps = [
     "libgfxstream_host_openglesdispatch",
     "libgfxstream_host_rendercontrol_dec",
     "libgfxstream_host_renderdoc",
-    "libgfxstream_host_snapshot",
     "libgfxstream_host_tracing",
     "libgfxstream_host_vulkan_cereal",
     "libgfxstream_host_vulkan_emulatedtextures",
@@ -116,7 +115,6 @@ cc_defaults {
         "libgfxstream_host_decoder_common",
         "libgfxstream_host_features",
         "libgfxstream_host_renderdoc",
-        "libgfxstream_host_snapshot",
         "libgfxstream_host_tracing",
         "libgfxstream_host_gles2_dec",
         "libgfxstream_host_glsnapshot",
@@ -196,7 +194,6 @@ cc_test_host {
         "libgfxstream_common_image",
         "libgfxstream_common_logging",
         "libgfxstream_common_testenv",
-        "libgfxstream_host_snapshot",
         "libgfxstream_host_test_support",
         "libgfxstream_host_vulkan_server",
         "libgfxstream_host_address_space",
