+++ host/common/include/gfxstream/host/lazy_snapshot_object.h
