+++ host/gl/glestranslator/common/BUILD.bazel
@@ -18,9 +18,9 @@ cc_library(
         "//common/etc:gfxstream_etc",
         "//common/logging:gfxstream_common_logging",
         "//host:gfxstream_backend_headers",
+        "//host/common:gfxstream_host_common",
         "//host/decoder_common:gfxstream_host_decoder_common",
         "//host/gl/OpenGLESDispatch:gfxstream_host_openglesdispatch_headers",
-        "//host/snapshot:gfxstream_host_snapshot",
         "//third_party/opengl:gfxstream_egl_headers",
         "//third_party/opengl:gfxstream_gles2_headers",
         "//third_party/opengl:gfxstream_gles3_headers",
