+++ host/gl/glestranslator/egl/meson.build
@@ -57,29 +57,28 @@ lib_egl_translator = static_library(
   objcpp_args: egl_cpp_args + gfxstream_host_args,
   dependencies: egl_deps,
   include_directories: [
-    inc_host_decoder_common,
     inc_common_base,
+    inc_common_logging,
     inc_common_utils,
     inc_etc,
     inc_gfxstream_server,
     inc_gl_common,
     inc_gl_openglesdispatch,
     inc_host_common,
+    inc_host_decoder_common,
     inc_host_library,
-    inc_common_logging,
-    inc_host_snapshot,
     inc_include,
     inc_opengl_headers,
     inc_stream_servers,
     inc_x11_headers,
   ],
   link_with: [
-    lib_host_decoder_common,
     lib_common_base,
+    lib_common_logging,
     lib_common_utils,
     lib_gl_common,
     lib_host_common,
+    lib_host_decoder_common,
     lib_host_library,
-    lib_common_logging,
   ],
 )
