+++ host/gl/glestranslator/gles_v2/meson.build
@@ -19,22 +19,21 @@ lib_glesv2_translator = static_library(
   cpp_args: gfxstream_host_args,
   include_directories: [
     inc_common_base,
+    inc_common_logging,
     inc_etc,
     inc_gfxstream_server,
     inc_gl_common,
     inc_gl_openglesdispatch,
     inc_host_common,
     inc_host_decoder_common,
-    inc_common_logging,
-    inc_host_snapshot,
     inc_include,
     inc_opengl_headers,
     inc_stream_servers,
   ],
   link_with: [
     lib_common_base,
+    lib_common_logging,
     lib_host_common,
     lib_host_decoder_common,
-    lib_common_logging,
   ],
 )
