--- host/snapshot/Android.bp
@@ -1,29 +1,0 @@
-// Copyright 2025 The Android Open Source Project
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either expresso or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package {
-    default_applicable_licenses: ["hardware_google_gfxstream_license"],
-}
-
-cc_library_static {
-    name: "libgfxstream_host_snapshot",
-    defaults: ["gfxstream_host_cc_defaults"],
-    header_libs: [
-        "libgfxstream_backend_headers",
-    ],
-    export_header_lib_headers: [
-        "libgfxstream_backend_headers",
-    ],
-    export_include_dirs: ["include"],
-}
