--- host/snapshot/BUILD.bazel
@@ -1,21 +1,0 @@
-load("@rules_cc//cc:defs.bzl", "cc_library")
-load("//:build_variables.bzl", "GFXSTREAM_HOST_COPTS", "GFXSTREAM_HOST_DEFINES", "GFXSTREAM_HOST_LINKOPTS")
-
-package(
-    default_applicable_licenses = ["//:gfxstream_license"],
-    default_visibility = ["//:gfxstream"],
-)
-
-cc_library(
-    name = "gfxstream_host_snapshot",
-    hdrs = [
-        "include/snapshot/LazySnapshotObj.h",
-    ],
-    copts = GFXSTREAM_HOST_COPTS,
-    defines = GFXSTREAM_HOST_DEFINES,
-    linkopts = GFXSTREAM_HOST_LINKOPTS,
-    strip_include_prefix = "include",
-    deps = [
-        "//host:gfxstream_backend_headers",
-    ],
-)
