--- host/snapshot/meson.build
@@ -1,4 +1,0 @@
-# Copyright 2025 Android Open Source Project
-# SPDX-License-Identifier: Apache-2.0
-
-inc_host_snapshot = include_directories('include')
