+++ host/testlibs/support/Android.bp
@@ -46,7 +46,6 @@ cc_test_library {
         "libgfxstream_common_logging",
         "libgfxstream_host_decoder_common",
         "libgfxstream_host_iostream",
-        "libgfxstream_host_snapshot",
         "libgfxstream_oswindow_test_support",
         "libgmock",
     ],
