+++ host/vulkan/CMakeLists.txt
@@ -66,7 +66,6 @@ target_link_libraries(gfxstream-vulkan-server PUBLIC
         gfxstream_host_iostream
         gfxstream_host_library
         gfxstream_host_renderdoc
-        gfxstream_host_snapshot.headers
         gfxstream_host_tracing
         gfxstream_openglesdispatch.headers
         gfxstream_vulkan_headers
